+++ net-snmp/net-snmp-5.2.1.2/agent/mibgroup/agent/nsDebug.c
@@ -1,3 +1,4 @@
+#include <stddef.h>
 #include <net-snmp/net-snmp-config.h>
 #include <net-snmp/net-snmp-includes.h>
 #include <net-snmp/agent/net-snmp-agent-includes.h>
@@ -290,7 +291,7 @@ get_first_debug_entry(void **loop_context, void **data_context,
 
     snmp_set_var_value(index, dbg_tokens[i].token_name,
 		       strlen(dbg_tokens[i].token_name));
-    *loop_context = (void*)i;
+    *loop_context = (void*)(ptrdiff_t)i;
     *data_context = (void*)&dbg_tokens[i];
     return index;
 }
@@ -300,7 +301,7 @@ get_next_debug_entry(void **loop_context, void **data_context,
                       netsnmp_variable_list *index,
                       netsnmp_iterator_info *data)
 {
-    int i = (int)*loop_context;
+    int i = (int)(ptrdiff_t)*loop_context;
 
     for (i++; i<debug_num_tokens; i++) {
         if (dbg_tokens[i].token_name)
@@ -311,7 +312,7 @@ get_next_debug_entry(void **loop_context, void **data_context,
 
     snmp_set_var_value(index, dbg_tokens[i].token_name,
 		       strlen(dbg_tokens[i].token_name));
-    *loop_context = (void*)i;
+    *loop_context = (void*)(ptrdiff_t)i;
     *data_context = (void*)&dbg_tokens[i];
     return index;
 }
